refactor(Property,-Advertisement-and-Booking-modules): Merge Property, Advertisement django apps into Booking app
