--- src/advertisement/domain/model.py
@@ -1,7 +1,0 @@
-class Advertisement:
-    def __init__(self, property, platform_name, platform_tax, created_at, updated_at):
-        self.property = property
-        self.platform_name = platform_name
-        self.platform_tax = platform_tax
-        self.created_at = created_at
-        self.updated_at = updated_at
