--- src/advertisement/serializers.py
@@ -1,7 +1,0 @@
-from rest_framework import serializers
-from advertisement.models import Advertisement
-
-class AdvertisementSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = Advertisement
-        fields = "__all__"
