+++ src/booking/serializers.py
@@ -1,7 +1,17 @@
 from rest_framework import serializers
-from booking.models import Booking
+from booking.models import Booking, Advertisement, Property
 
 class BookingSerializer(serializers.ModelSerializer):
     class Meta:
         model = Booking
+        fields = "__all__"
+
+class AdvertisementSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = Advertisement
+        fields = "__all__"
+
+class PropertySerializer(serializers.ModelSerializer):
+    class Meta:
+        model = Property
         fields = "__all__"
