--- src/property/serializers.py
@@ -1,7 +1,0 @@
-from rest_framework import serializers
-from property.models import Property
-
-class PropertySerializer(serializers.ModelSerializer):
-    class Meta:
-        model = Property
-        fields = "__all__"
