--- src/property/urls.py
@@ -1,7 +1,0 @@
-from django.urls import path
-from property import views
-
-urlpatterns = [
-    path('', views.PropertyList.as_view()),
-    path('<uuid:pk>/', views.PropertyDetail.as_view())
-]
