+++ src/real_state_rest_api/settings.py
@@ -41,8 +41,6 @@ INSTALLED_APPS = [
     'django.contrib.messages',
     'django.contrib.staticfiles',
     'rest_framework',
-    'property',
-    'advertisement',
     'booking'
 ]
