--- tests/test_real_state_rest_api.py
@@ -1,5 +1,0 @@
-from real_state_rest_api import __version__
-
-
-def test_version():
-    assert __version__ == '0.1.0'
